+++ packages/api/src/app.ts
@@ -3,13 +3,16 @@ import fastifyCors from 'fastify-cors'
 import fastifyStatic from 'fastify-static'
 import { resolve } from 'path'
 
+import { env } from './_services/env/index.js'
 import { routesActivityPatterns } from './activity-patterns/index.js'
 import { routesIndex } from './index/index.js'
 import { routesProjectSite } from './projects-and-sites/index.js'
+import { routesSpecies } from './species/index.js'
+import { routesIucn } from './species/iucn/index.js'
 import { routesSpeciesRichness } from './species-richness/index.js'
 
 export const app = fastify({
-  logger: process.env.NODE_ENV !== 'production'
+  logger: env.NODE_ENV !== 'production'
 })
 
 // Register plugins
@@ -17,5 +20,5 @@ await app.register(fastifyCors)
 await app.register(fastifyStatic, { root: resolve('./public') })
 
 // Register routes
-const routePlugins = [routesIndex, routesProjectSite, routesSpeciesRichness, routesActivityPatterns]
+const routePlugins = [routesIndex, routesProjectSite, routesSpecies, routesSpeciesRichness, routesActivityPatterns, routesIucn]
 await Promise.all(routePlugins.map(plugin => app.register(plugin)))