+++ packages/website/src/_services/feature-toggles/index.ts
@@ -1,0 +1,15 @@
+import { mapValues } from 'lodash'
+
+const featureTogglesRaw = {
+  /* START: Declare toggles here */
+
+  // Example:
+  // toggleAoLine: import.meta.env.VITE_TOGGLE_AO_LINE
+
+  /* STOP: Declare toggles here */
+}
+
+// Setup types & export
+type FeatureTogglesRaw = typeof featureTogglesRaw
+export type FeatureToggles = { [K in keyof FeatureTogglesRaw]: boolean}
+export const featureToggles: FeatureToggles = mapValues(featureTogglesRaw, value => value as boolean)